Turned GallerySlider into component
