+++ bloom-and-vine/src/Components/gallery-display/galleryDisplay.css
@@ -36,5 +36,5 @@
 
 .inactive {
     font-weight: 500;
-    opacity: 0.5;
+    opacity: 0.35;
 }
